+++ warpgate-common-http/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod auth;
 pub mod ext;
 pub mod logging;
+mod request;
 
 pub use auth::{AuthenticatedRequestContext, RequestAuthorization, SessionAuthorization};
 use poem::http::HeaderName;
