+++ warpgate-common-http/src/logging.rs
@@ -5,26 +5,9 @@ use poem::http::{Method, StatusCode, Uri};
 use poem::web::RemoteAddr;
 use poem::{Addr, Request};
 use tracing::*;
-use warpgate_common::http_headers::X_FORWARDED_FOR;
 use warpgate_core::{Services, WarpgateServerHandle};
 
-use crate::auth::first_forwarded_header_value;
-
-fn trusted_client_ip(
-    req: &Request,
-    remote_ip: Option<String>,
-    trust_x_forwarded: bool,
-) -> Option<String> {
-    if trust_x_forwarded
-        && let Some(ip) = req
-            .header(&X_FORWARDED_FOR)
-            .and_then(first_forwarded_header_value)
-    {
-        Some(ip.to_string())
-    } else {
-        remote_ip
-    }
-}
+use crate::request::trusted_client_ip;
 
 pub async fn get_client_ip(req: &Request, services: &Services) -> Option<String> {
     let trust_x_forwarded_headers = {
@@ -83,45 +66,4 @@ pub fn log_request_result(method: &Method, url: &Uri, client_ip: Option<&str>, s
 pub fn log_request_error<E: Debug>(method: &Method, url: &Uri, client_ip: Option<&str>, error: &E) {
     let client_ip = client_ip.unwrap_or("<unknown>");
     error!(%method, %url, ?error, %client_ip, "Request failed");
-}
-
-#[cfg(test)]
-mod tests {
-    use poem::Request;
-
-    use super::*;
-
-    #[test]
-    fn trusted_client_ip_uses_first_forwarded_for_value() {
-        let req = Request::builder()
-            .header(&X_FORWARDED_FOR, "203.0.113.10, 10.0.0.2")
-            .finish();
-
-        assert_eq!(
-            trusted_client_ip(&req, Some("10.0.0.1".to_string()), true),
-            Some("203.0.113.10".to_string())
-        );
-    }
-
-    #[test]
-    fn trusted_client_ip_falls_back_when_forwarded_for_is_empty() {
-        let req = Request::builder().header(&X_FORWARDED_FOR, " , ").finish();
-
-        assert_eq!(
-            trusted_client_ip(&req, Some("10.0.0.1".to_string()), true),
-            Some("10.0.0.1".to_string())
-        );
-    }
-
-    #[test]
-    fn client_ip_ignores_forwarded_for_when_not_trusted() {
-        let req = Request::builder()
-            .header(&X_FORWARDED_FOR, "203.0.113.10")
-            .finish();
-
-        assert_eq!(
-            trusted_client_ip(&req, Some("10.0.0.1".to_string()), false),
-            Some("10.0.0.1".to_string())
-        );
-    }
 }
